+++ pkg/helm/tune.go
@@ -3,6 +3,7 @@ package helm
 import (
 	"embed"
 
+	"github.com/blang/semver/v4"
 	"groundcover.com/pkg/k8s"
 	"k8s.io/apimachinery/pkg/api/resource"
 )
@@ -16,6 +17,13 @@ const (
 	AGENT_DEFAULT_CPU_THRESHOLD    = "1000m"
 	AGENT_DEFAULT_MEMORY_THRESHOLD = "1024Mi"
 	AGENT_LOW_RESOURCES_PATH       = "presets/agent/low-resources.yaml"
+
+	// Starting from Linux kernel version 5.11, eBPF maps are accounted for in the memory cgroup
+	// of the process that created them. For this reason we need to increase the memory limit for
+	// the agent.
+	// https://github.com/cilium/ebpf/blob/v0.16.0/docs/ebpf/concepts/rlimit.md#resource-limits
+	AGENT_KERNEL_5_11_PRESET_PATH = "presets/agent/kernel-5-11.yaml"
+	KERNEL_5_11_SEMVER_EXPRESSION = ">=5.11.0"
 
 	EMPTYDIR_STORAGE_PATH = "presets/backend/emptydir-storage.yaml"
 
@@ -39,7 +47,7 @@ type AllocatableResources struct {
 	NodeCount   int
 }
 
-func GetAgentResourcePresetPath(allocatableResources *AllocatableResources) string {
+func GetAgentResourcePresetPath(allocatableResources *AllocatableResources, maxKernelVersion semver.Version) string {
 	defaultCpuThreshold := resource.MustParse(AGENT_DEFAULT_CPU_THRESHOLD)
 	defaultMemoryThreshold := resource.MustParse(AGENT_DEFAULT_MEMORY_THRESHOLD)
 
@@ -48,6 +56,10 @@ func GetAgentResourcePresetPath(allocatableResources *AllocatableResources) stri
 
 	if minAllocatableCpu <= defaultCpuThreshold.AsApproximateFloat64() || minAllocatableMemory <= defaultMemoryThreshold.AsApproximateFloat64() {
 		return AGENT_LOW_RESOURCES_PATH
+	}
+
+	if semver.MustParseRange(KERNEL_5_11_SEMVER_EXPRESSION)(maxKernelVersion) {
+		return AGENT_KERNEL_5_11_PRESET_PATH
 	}
 
 	return DEFAULT_PRESET